Comments are now bold.
The `reverse` hack I used to use for the light background is pretty brittle and would often cake half the screen grey because of leading whitespace.
Grey comments on a white background are unreadable, so they're black.

I believe it's been fixed recently, but the instances of Gnu Screen I encounter in the real world still can't grok italics.
I'm ok with minor elements such as folds being displayed wrong because someone didn't read the spec properly back in 1987, but I simply can't have that for something as important as comments.